Several updates to ClockView.swift
I was really disappointed with how I wrote the code.  The part that bugged me the most was the incrementByOneSecond() function, which I knew could be done in a more elegant way.  I found a function that my teacher showed me in my C++ class and used that.  I also thought that putting seconds, minutes, and hours into an array and making them timeIntervals[0], timeIntervals[1], and timeIntervals[2], respectively, just so I could use a for loop was terrible for readability, so I went back to making seconds, minutes, and hours three separate variables, and I made a separate drawHands function, which I called three times in the drawRect function.